Updated openstack/openstack Project: openstack/horizon  bda27b02d6514ef96fc50920b6282b7d5ac965cc
"QoS spec" page is displayed as a modal dialog

This page is implemented as a modal dialog.
Like "View Extra Specs" page. But this page is displayed as
a static page. This patch modifies it.

In addition this, I addressed to redirect to main non-modal index
page (it is Volume Type page) after processing create/update/delete
done, because if destination of redirect is modal page,
it can not redirect correctly (displayed with non-modal).

Change-Id: Ib1c0a2c11a24db46f1d36a719d3881122968d102
Closes-Bug: #1526700