Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  41e9678d6e3a33fbd38818b93a5f66ee544fdc1b
Add separate linkcheck env and allow dev to select builder

Add separate linkcheck tox environment to be used in non-voting Jenkins
job to watch for stale and broken links in the sphinx generated html
documentation. Include a small addition to make it easier for
contributors to select individual builders when executing the docs tox
environment.

Add minimum dependency of pbr 1.0.0, as that was the first version to
allow passing of sphinx builders on the command line to override the
internal defaults or those set in the configuration file.

This will make it easier to spot problems with just the links, and also
ensure that when the sphinx project releases with proper warniserror fix
for the linkcheck builder, that html and man page documentation can
continue to be built even if a previous valid link has become stale or
is temporarily unreachable.

Change-Id: I80446abe96824a444a04cc526df6cc38632a4607