Updated openstack/openstack Project: openstack/oslo.versionedobjects  37005c310eb8a6eb87f455adbcc93cbed7dccbb5
Add temporary registry pattern to VersionedObjectRegistry

There is a temp registry pattern [1] where you can backup the object
registry, register a class locally, and then restore the original
registry. This could be used for test objects that do not need to be
registered permanently but will have calls which lookup registration.
The FakeResource object in patch [2] is such a use case.

This change is to wrap the pattern rather than accessing
oslo.versionedobjects library internals.

[1] https://review.openstack.org/#/c/243826/4/cinder/test.py
[2] https://review.openstack.org/#/c/258026/

Change-Id: I02036becb32f8d3d11170c71e14e0a2db55d1f76