Updated openstack/openstack Project: openstack-dev/hacking  32084595946f3ddd985e36d35bea022958716b3b
Avoid git clone in integration test

Allow script to be called with a third argument which is a path to a
checked out directory. This allows setting up the repo externally using
zuul-cloner in the infra script and allows usage of depends-on.

Simplify the script a bit, we can remove the special handling for file
and also do not need to give a default argument, this is called from
tox.

Change-Id: I3b6180b6b274966de63fe281021ff9f4186a929b
Needed-By: I828c1875e5f008fa55b693ad210421794395f623