Updated openstack/openstack Project: openstack/keystonemiddleware  63f0fc3e61a4407e419324c24ae619cc22d87545
Merge pep8 and bandit into linters

Create new linters environment consisting of pep8 and bandit.
Keep pep8 as an alias for linters and add a message when invoked.

Once this change is in, the infra job can be changed to only run
linters job, this allows to run one less job for each change.

Change-Id: Idff7d7134c50aa4740ed87eeeda83f7cca32f3db