Updated openstack/openstack Project: openstack/python-keystoneclient  0a0419d2b8a5580fbdce9a673b132af5fa0c4d57
Merge pep8 and bandit into linters

Create new linters environment consisting of pep8 and bandit.
Keep pep8 as an alias for linters and add a message when invoked.

Once this change is in, the infra job can be changed to only run
linters job, this allows to run one less job for each change.

Change-Id: Ib15be94491232d63a345d6fc19869095a2537310