Updated openstack/openstack Project: openstack-infra/nodepool  b8e432ebd6d4805005af7b36deff8664626d2e3d
Remove workaround for early paramiko import

As discussed in the original bug [1] one of the reasons to not
daemonize after import of paramiko is that it creates a global
Crypto.Random() object which is used around the place, however the
close of fd's in the daemonize breaks that object's connection to
/dev/urandom.

However, it seems that this issue is no longer as paramiko now
directly uses urandom and has no global object [2].  Requirements is
updated to the first release after this change.

We have also removed the use of the global statsd object in
I6d339a8c631f8508a60e9ef890173780157adefd.

Thus we should be able to return this to it's simpler form

[1] https://github.com/paramiko/paramiko/issues/59
[2] https://github.com/paramiko/paramiko/commit/6f211115f49edcea7d23b764d7cf3a84ff12f5f0

Change-Id: I6cf3ba0df857f4213b0ec658ebd0511a24618175