Updated openstack/openstack Project: openstack/glance_store  8a72e9c516e5173fd609bcc476b911d033038472
Document supported drivers and maintainers

As per the latest communication sent to openstack-dev[0], this commit
adds the list of drivers and maintainers to the glance_store
documentation. The drivers listed here are considered supported by the
overall community but mainly by the listed maintainers.

Drivers that currently exist in the source tree that are not listed here
will be marked as deprecated following what was communicated in the
aforementioned thread[0].

[0] http://lists.openstack.org/pipermail/openstack-dev/2015-December/081966.html

Change-Id: I0ae750a8bbd936e421a5221b7a97fef1f0d02750