Updated openstack/openstack Project: openstack/glance_store  3a135e613073664ddf51c05c8788e089f0e532b4
Deprecate the S3 driver

As mentioned in this mailing list thread[0], we've started a cleanup
process for the drivers in glance_store. In order to guarantee some
level of stability, we need each driver to have an assigned maintainer
that would be the main contact point for things related to such driver.

The S3 driver doesn't have one. Therefore, as advised in the ML thread,
we're marking it as deprecated and prepare it for removal during the N
development cycle.

[0] http://lists.openstack.org/pipermail/openstack-dev/2015-December/081966.html

Change-Id: I139c36c2d591a55f0f91ea32efa963f80c28e444