Updated openstack/openstack Project: openstack/nova-specs  8fd058f7c4762e2c705a4556b07242d7e0434efb
REST API changes for user settable server description

This patches adds changes to the Nova REST API to allow users
to create a server with a description, rebuild a server with
description, update the description,
and get the description in the server details.

Note: Future commits will be done to support the server description
in python-novaclient and openstack-client.

APIImpact

Implements blueprint: user-settable-server-description

Change-Id: I262e1fd5cfcc6f78d7979c122f00434e95e3d709