Updated openstack/openstack Project: openstack-infra/project-config  d150cb26a3a3276e895af342a5230d4a186c2f94
Define criteria for skipping multinode-live-migration

gate-tempest-dsvm-multinode-live-migration job should be
executed if changes were done in its post_test_hook.
Prior to this patch such changes was not tested, as
devstack long-running jobs are skipped if changes are
made in tests directory of nova project. This patch
redefines skip criterias for live-migration job.

Change-Id: Idee4bf82abc496b21077d39e0e2eaf03f100859b