Updated openstack/openstack Project: openstack-infra/project-config  6f4953fa609a8efd36c3b33ccbfaeaab6e7391be
Add exclusiveGroupPermissions to sandbox's acl

Currently the gitweb access information lists sandbox's acl
as giving +2 +A to all registered users. Looking at a patch to
sandbox reveals that +2 +A is not in fact a permission all
registered users can access.

With the belief that since the all projects acl setting gives
+1 to all registered users and therefore that the project
specific sandbox acl is not overridding the all projects acl,
the patch adds exclusiveGroupPermissions to the sandbox acl
to override the all projects acl and grant +2 +A to all
registered users for the sandbox repo.

Change-Id: I4f69d9ec3fdcdc524760f153fa61b0d9108cd8c6