Updated openstack/openstack Project: openstack/keystone  fbf4eccb2949cd1db13fcb2a931d4903fb02a502
Add support for strict url safe option on new projects and domains

Building on the earlier patch that provdided the 'new' url name
restriction, this patch adds the 'strict' open that prevents
authenticating to projects and domains with unsafe names.

A release note and config documentation is also added that covers
both this and the earlier patch.

Partially Implements: blueprint url-safe-naming

Change-Id: Ie69025e7759bae1067e05d9190bede192a5e6830