Updated openstack/openstack Project: openstack/neutron-lbaas  860d7b73b315227cf993cb81ac5ed66e964bd4e4
Refactor BarbicanAuth to allow for configurable auth method

There is now a new configuration option "barbican_auth" in the
certificates section, to specify which auth plugin to use when
communicating with Barbican. This is because the default option (using
ACLs inside Barbican to control access) should be ok as a default
workflow, but it might be required to use other methods depending on
your deployment. For example, another possible auth method would be
BarbicanTrustAuth, utilizing Keystone Trusts.

While we're in this file, prune the unused (and really bad) method that
would *actually* delete certs from Barbican (not in our scope).

This is (nearly) a direct port of the work done in:
    https://review.openstack.org/#/c/216140/

Change-Id: Id8386c36c7acfd954bc2309947345dfe31124b16