Updated openstack/openstack Project: openstack/cinder  789c81861f43323f797053a94021b4aef333b7f6
Activate sparse copy for Netapp

Activating sparse copy of files since using
"volume migration" or "retype" the size of the resulting
file is larger than before. Due to the reason that the
sparse blocks in the file are copied over completely. This
is inefficient performance wise and wastes space on the
storage itself.

Closes bug: #1534649

Change-Id: I3be7baf0255d838e6f5e16302c85d318061f4004