Updated openstack/openstack Project: openstack-infra/system-config  35dd62572792404e7b80695c9c23f1395aae0d29
Add wheel_mirror_slaves that use pypi-mirror

We want to start building wheels for as many of our dependencies as
possible in order to speed up test runs. Resurrecting parts of the old
pypi-mirror code.

We have pre-seeded this with a group of common packages, taken roughly
from the bindep output from global-requirements.

Change-Id: Ie33861dbcb413c34012820fda76f8be94fb1d151