Updated openstack/openstack Project: openstack/nova  3b2c87e25924de8d5c5b00a7ed71f4efa3893d1c
Add note to ComputeNode.numa_topology

As the ComputeNode is being converted away from a dict, it was
discovered that numa_topology is held in the ComputeNode as a
StringField, not an ObjectField containing a NUMATopology. Because
changing types within minor versions of an object is not a good thing to
do, a note is added to change the field over to use the NUMATopology
object in the next major version of the ComputeNode object (2.0).

Change-Id: Ife9f7aefbe64638521abcd27a9e182612fab46c7