Updated openstack/openstack Project: openstack/nova-specs  763a287fabddce5d437d07b66a9b45b8e220487c
pbr > 1.6 to avoid pip 8 distutils issue

"Detected a distutils installed project" issue is hitting
for argparse for specs repo.
In case of nova specs repo it comes from testtools.

As a workaround, making pbr version > 1.0 so that venv having pip
< 8 and works fine. pbr < 1.0 endup with pip 8 and fails.

pbr < 1.0 has pip explicitly in requirement.txt and trigger
to update the pip to 8.

pbr > 1.0 does not have the requirement.txt thing and does not
trigger pip updates for venv.

This is issue with all specs repo having distutils installed projects
because g-r requirement update does not sync the specs repo requirements.
But that is another thing can be fixed if there is no reason of not
updating specs requirements.

Change-Id: I31a1353d161e5e5687332cd8300ba5dd4f992a67