Updated openstack/openstack Project: openstack/neutron  03b5caf78a92e426508a363b08e5274a47d92435
Fixed qos devstack service name (should be q-qos) for -plus gate hook

This hook was introduced for experimental -plus job and is supposed to
enable advanced features like qos. Sadly, the service name was not
correct, so the service was not enabled.

The patch that introduced the hook:
I9f73b938539586022aeb22e3a32877816e1dcaba

Change-Id: I1c2e758341825d031a473d837e9a2e707e5ce5f3