Updated openstack/openstack Project: openstack/oslo.utils  123e12201aebeeccc22e711eca40fa4cc3e13b2f
Create secretutils and include 'constant_time_compare' function

This code (or a version of it) is being shared by at least nova
and keystonemiddleware and it seems like a good idea to move it
to being common shared code (especially due to the importance
of getting this code correct).

This adds an initial secretutils and adds tests for it.

Change-Id: Ia603202a065d5b345608e712f63f7af21fd74dea