Updated openstack/openstack Project: openstack/oslo.utils  387c1068b6b3cc51795d36f382bb996c5a2a0a1d
Add encodeutils.to_utf8() function

The function replaces a very common pattern in code base supporting
Python 2 and Python 3:

    if isinstance(text, six.text_type):
        text = text.encode('utf-8')

to_utf8() accepts subtypes of bytes and six.text_type. For example,
oslo.i18n Message objects are accepted and encoded to UTF-8 as
expected.

Using encodeutils.safe_encode(text) is not reliable because it relies
on the current locale encoding which can be ASCII, whereas UTF-8 is
expected. Having to write encodeutils.safe_encode(text,
incoding='utf-8') is not obvious and error-prone (it's easy to forget
the incoming parameter).

Change-Id: I00463716b6012cbef383855999f63f99f2f52540