Updated openstack/openstack Project: openstack/trove-integration  e395112ada42495fee6ce0c4a8884cd34e422cca
Adds the dib elements before the triplo elements

There is a duplicate package of pip-and-vitrualenv in dib elements as
well as in triple-image-elements. Adding the dib elements before the
tripleo elements in the ELEMENTS_PATH environment variable will allow
dib to find its own element first instead of the of tripleo
element. The tripleo element for pip-and-virtualenv need an update to
work right. This will make sure that we are using the common set of
elements before the tripleo elements in the future.

Change-Id: I7069fcab3d16a960dc52757c087907ee8c41b309
Closes-Bug: #1536530