Updated openstack/openstack Project: openstack/python-openstackclient  f51f26868c3eb8fd0a4e624ea3066689e7f6b8a7
log_method: get logger from decorated method if unspecified

This commit makes 'log' optional.
'log' attribute of each command class does not exist
when the class is defined because 'log' is now setup
dynamically when a class is instantiated. Instead log_method
looks for a logger from a decorating method.

compute.v2.server is changed in this commit as an example.

Change-Id: Ic4d128f8e027d3b8e6f884f31369e9085c0f0871
Partial-Bug: #1532294