Updated openstack/openstack Project: openstack-infra/project-config  03916cc92750fe8d71f4384c03b73ba6b417613c
fix ceilometer integration job to handle slow gate

ceilometer tries to connect to gnocchi when started. a retry mechanism
is in place if gnocchi is unavailable. on a slow gate, the retry
limit is hit before gnocchi is started so ceilometer abandons its
attempt to connect and resulting checks fail as no data is recorded.

this patch changes it so gnocchi is started first to ensure it's ready
when ceilometer attempts to connect to it.

Change-Id: I77bc50235222422e5a0f7b6797ae0ed8640a3495