Updated openstack/openstack Project: openstack/python-neutronclient  db3e862265ed031f911ce80cc554a6c5f397bd90
refactor: Get rid of usage of get_data in favor of take_action

Previously the dropped OpenStackCommand uses get_data()
as an interface which subclasses should implement.
On the other hand, cliff Command recommends to implement
take_action() in subclasses and there is no reason not to
follow this convention.

Partial-Bug: #1532258
Change-Id: I164d6f31152df621611985a84eef76e7dfa54d44