Updated openstack/openstack Project: openstack-infra/storyboard-webclient  68213f1f8db88de3c884393b7b4baa387c5d59ca
Change Hover Cursor for Interactable Titles

This makes the cursor change to a pointer when hovering over the
interactive titles. These are the ones that allow sorting projects,
stories and project-groups by 'created', 'updated', etc. Previously,
the cursor changed to a 'text' type cursor, which was misleading.

Story: https://storyboard.openstack.org/#!/story/2000061

Change-Id: I9670dd273d87e0eb5f7d73d7fec50a3f9b83215d