Updated openstack/openstack Project: openstack/nova  2c59e613c04ca373e298972f421749def5d4236e
VMware: improve instance names on VC

Up until now the instance name of the VM on the VC was the UUID of
the instance. This is very confusing for administrators and they
would like some context for the VM.

This will now be:
 <display-name> (<instance-uuid>)

Once the the instance is created we will perform a rename. The reason
for this is that the disks are all stored under a folder on the datastore
that has the name of the instance UUID.

NOTE: this is only for the display on the VC. The search for the VM's
      is done via the instanceUUID configured on the VM. If that fails
      then the nvp.uuid extra spec is used for the search.

This completed the blueprint vmware-better-display-names

Closes-bug: #1330873

Change-Id: Ibeb000f4c1d88fa296390143b0cf98e05e735086