Updated openstack/openstack Project: openstack/tripleo-heat-templates  fe2133fa8bfface39bdf132b55b692ac349a85e1
Add update yaml backward compatibe with PublicVirtualIP on ctlplane

In previous releases, when not using network isolation, we used to create
two different VIPs for the ControlVirtualIP and the PublicVirtualIP both on
the ctlplane network. Later we moved into a configuration with a single
VIP instead so we need a compatibility yaml for those updating from old
versions which preserves both the IPs; one of the two is deleted
otherwise.

Also updates README.md with a short description of the use case.

Change-Id: Iae08b938a255bf563d3df2fdc0748944a9868f8e