Updated openstack/openstack Project: openstack/glance_store  30ae6b020c9a903d50604d5b66a490dfe6de85b0
Add signature verifier to backend drivers

In order to use the 'sign-the-data' approach instead of the 'sign-the-
hash' approach for signature verification, the verifier must be
updated with the image data at the same time that the image data is
being read in order to update the checksum.

This patch adds the ability for glance to pass in a verifier object to
the backend drivers when an image is being uploaded, which is given
the image data by the backend drivers, and used to verify the
signature in glance.

The following drivers are supported:
* file
* rbd
* sheepdog
* s3
* vsphere

The swift driver is supported for images that are greater than the
'large_object_size' since glance_store does not read the data for
images that are less than 'large_object_size' (swiftclient does the
reading), and as a result the verifier cannot be updated with the
data.  A child patch addresses this issue.

The cinder and http drivers are not supported because they do not have
the 'write' capability.

Partial-Bug: #1516031
Change-Id: I43799e6a4a6643a23769af8d839a2beb4e0ff9bf