Updated openstack/openstack Project: openstack/manila  1461eb7e4d6f63590374608c1388d8fbd7872be3
Hitachi HNAS driver refactoring

Hitachi HNAS driver refactoring in order to improve modularization.
With this patch, the code used for communicating with HNAS is
separated from the driver, so it is easier to change the
communication protocol. This patch also fixes some minor bugs from
previous versions of the driver.

DocImpact
Change-Id: Iffdd6ea863dea5ad73c3cb1cdd9cdc0b75d695f1
Implements: blueprint hitachi-hnas-driver-refactoring