Updated openstack/openstack Project: openstack/cinder  e2c06b9c85e81013316a4e57414d7174b53d70aa
NetApp FC drivers should not set 'rw' access mode

The volume manager is responsible for setting the initiator access
mode in initialize_connection based on volume metadata.  The NetApp
FC drivers currently hardcode the access mode to 'rw'.  So the fix
is to stop setting the access_mode value in the connection info that
is returned from the NetApp drivers.

Partial-Bug: 1526152
Change-Id: I587bba62ebfc067bb362ce95b324b4a74c0e02d0