Updated openstack/openstack Project: openstack/glance  2e2c8ff0cbb043ec9f14ffe4fb1e990c8be759c3
Fix re-adding deleted members to an image in v1

If you replace the membership list of an image with some members
that were previously deleted, we fail to re-add those deleted
members.

This seems to be a consequence of a previous commit
d6800e143ddff8898d693bbef65d2143544d7ad0 which failed to update
the deleted records.

Closes-Bug: 1527143

Change-Id: Ic3ca9b56712a99652e65cb0a4e3f1e0ba15b8593