Updated openstack/openstack Project: openstack/diskimage-builder  d8b3acf816469784f2db7a9704940317085dfdad
Force dib-python symlink creation

The undercloud actually has dib run twice on it - once to create
the instack image, and again when we run instack itself.  The
first run creates the dib-python symlink, and the second blows up
because the link already exists.  Force the link creation so the
script is idempotent.

Change-Id: I78f9e6f5afcf8ebe6d7911a7a434525ba7c737cf