Updated openstack/openstack Project: openstack/horizon  d4c4efbcce95b0f1d0334f97e30ad8d734e79117
Tolerate unconfigured glance service when loading image name

When loading the instances page, the glance service is used as a
fallback to obtain the image name for each instance. This patch adds
tolerance for an unconfigured glance service endpoint for the current
region so that the page will still load.

Change-Id: Ia512269ae4b069db85a1f13530856b9c7beafee3
Closes-Bug: #1361298