Updated openstack/openstack Project: openstack/horizon  008a2e665fedf7b0f170aa8196274f70bef4366c
In Stack Details do not display as links events leading to nowhere

Events that correspond to stack operations have the same physical id
as the stack has itself (and the same name), so the link which is
meant to point to some resource description, points to nowhere for
them. Better not render such links at all.

Change-Id: I2cff8efacbc00bed2989e5d4afee0690e712ad66
Closes-Bug: #1480997