Updated openstack/openstack Project: openstack-infra/project-config  d18dd6f65cd6c6faf4e5b5b447527e39ea79c56e
Reduce internap max servers from 72 to 64

We are seeing race condition in the Nova resource tracker when reaching
maximum compute capacity. Our hypothesis is that some instances are getting
scheduled to an hypervisor which doesn't have enough free resources
according to the resource tracker, rejecting the instance creation.

We suggest reducing the max servers to avoid hitting this race condition and
reducing the number of capacity alerts sent to our operation team.

Change-Id: Iff197a94ad988fb070efaeb6bb393bfaba6e01fb