Updated openstack/openstack Project: openstack/trove  0059578f2f16ebb04cea6c5815f774cc5252ae26
Time to get rid of most vestiges of slave_of

Commit https://review.openstack.org/#/c/245738/ eliminated slave_of in
the python-troveclient. The controller has long understood how to deal
with replica_of but there are a couple of tests that are still sending
in slave_of.

This commit removes slave_of from the tests and from the Trove public
API. The python-troveclient commit will depend on this commit as the
DSVM test(s) which are in the trove repository currently fail in that
change.

This commit also contains a change to tests/api/replication.py and
disables the test 'test_detach_replica' and
test_slave_is_not_read_only with a distinct SkipTest. This is to break
the deadly embrace between this change set and the related changeset
on the client side as circular dependencies across projects are not
allowed.

Change-Id: I87c8a14c06fa5df60ea3ed6b09412dc32fafcec7
Partial-Bug: #1461196
Related: I9b905716b704ea9ec9332b48b05c49cccbd04c8b