Updated openstack/openstack Project: openstack/governance  4e38b0efdb72b894a1944afb6b1cec77a8c03af3
Remove python-kollaclient from Kolla deliverables

As announced on the mailing list, the developers that contributed
the CLI codebase have decided to abandon their work and wait for
a fully integrated Ansible + Mesos solution before committing to
any further CLI development.  As such we are not merging this
repository and removing it as a deliverable.

Change-Id: I63437a8a31f54e63c041f591f9bcc98e0c70ce91